# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wine2u API",
        "description": " description",
        "contact": {
            "email": "dawud@effectstudios.co"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/user/forgotten-password/email": {
            "post": {
                "tags": [
                    "Forgot Password(Customer)"
                ],
                "summary": "Forgot Password",
                "description": "Forgot Password",
                "operationId": "authCustomerForgotPassword",
                "requestBody": {
                    "description": "Pass Your Email Address",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "example@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong input response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/password/reset": {
            "post": {
                "tags": [
                    "Forgot Password(Customer)"
                ],
                "summary": "Customer Password Reset",
                "description": "Customer Password Reset",
                "operationId": "authCustomerPasswordReset",
                "requestBody": {
                    "description": "Pass Your Email and New passwword Address",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "example@mail.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "token"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "yoursecret"
                                    },
                                    "password_confirmation": {
                                        "type": "password",
                                        "example": "yoursecret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong input response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}